Disabled stick commands when in CLI
This disabled stick commands when in CLI. Ideally they should be disabled all the time when connected to configurator. An issue has been discovered where the accelerometer can be trimmed via stick commands while in configurator. If this happens too much, it will cause a hang, and wipe all settings.
For testing, I added a `SYSTEM_MESSAGE` to the OSD to show when settings had been saved. This is pretty useful, so it may as well stay. Maybe this will reduce batteries being pulled out too early.