Add PINIO 3 & 4
There are plenty of flight controllers around now that use PINIO 1 and 2 internally for camera switching and switchable VTx power. This left no scope for user definable PINIO. A friend wanted to control a Caddx Peanut on such an FC. So I have implemented USER 3 and USER 4 PINIO options. Also in the PR is the target that was created for my friend.